+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/singularityhub/singularity-hpc/tree/master) (0.0.x)
+ - more packages and shell for a container (0.0.19)
  - adding description to docgen, and descriptions for all containers (0.0.18)
  - descriptions are now required
  - adding docgen command (0.0.17)